+++ src/csrs/pages/templates/editable/edit_path.py
@@ -3,7 +3,7 @@ from jinja2 import Environment
 
 from .... import schemas
 from ...loader import ENV
-from ..form_parts import EditableStr, EditableStrLong
+from ..forms import EditableStr, EditableStrLong
 
 
 class EditablePaths:
