+++ src/csrs/pages/templates/editable/edit_run.py
@@ -3,7 +3,7 @@ from jinja2 import Environment
 
 from .... import schemas
 from ...loader import ENV
-from ..form_parts import EditableStr, EditableStrLong, EditableSwitch
+from ..forms import EditableStr, EditableStrLong, EditableSwitch
 
 
 class EditableRuns:
