+++ src/csrs/pages/templates/forms/create_attr_str.jinja
@@ -1,4 +1,4 @@
-{% extends "templates/form_parts/_editable_attr.jinja" %}
+{% extends "templates/forms/_editable_attr.jinja" %}
 {% block editable %}
 <input 
     type="text"
