+++ src/csrs/pages/templates/forms/editable_attr_selection_group.jinja
