+++ src/csrs/pages/templates/new/_new.jinja
@@ -1,5 +1,5 @@
 {# defined: id, request #}
-{% import 'templates/form_parts/buttons.jinja' as buttons %}
+{% import 'templates/forms/buttons.jinja' as buttons %}
 
 <div class="accordion-item border-success">
     <!-- Header  -->
